+++ R/sim2data.r
@@ -3,19 +3,21 @@
 #' @export
 sim2data <- function(sim, to, use_saved_states=sim$save_states=="all",
                      overlap=FALSE, target_event=NULL,
-                     keep_only_first=FALSE, as_data_frame=FALSE,
-                     check_inputs=TRUE, ...) {
+                     keep_only_first=FALSE, remove_not_at_risk=FALSE,
+                     as_data_frame=FALSE, check_inputs=TRUE, ...) {
 
   if (check_inputs) {
     check_inputs_sim2data(sim=sim, use_saved_states=use_saved_states, to=to,
                           overlap=overlap, target_event=target_event,
-                          keep_only_first=keep_only_first)
+                          keep_only_first=keep_only_first,
+                          remove_not_at_risk=remove_not_at_risk)
   }
 
   if (to=="start_stop") {
     data <- sim2start_stop(sim=sim, use_saved_states=use_saved_states,
                            overlap=overlap, target_event=target_event,
-                           keep_only_first=keep_only_first)
+                           keep_only_first=keep_only_first,
+                           remove_not_at_risk=remove_not_at_risk)
   } else if (to=="long") {
     data <- sim2long(sim=sim, use_saved_states=use_saved_states)
   } else if (to=="wide") {
@@ -35,12 +37,14 @@ sim2data <- function(sim, to, use_saved_states=sim$save_states=="all",
 as.data.table.simDT <- function(x, keep.rownames=FALSE, to,
                                 overlap=FALSE, target_event=NULL,
                                 keep_only_first=FALSE,
+                                remove_not_at_risk=FALSE,
                                 use_saved_states=x$save_states=="all",
                                 check_inputs=TRUE, ...) {
   out <- sim2data(sim=x, to=to, use_saved_states=use_saved_states,
                   as_data_frame=FALSE, check_inputs=check_inputs,
                   overlap=overlap, target_event=target_event,
-                  keep_only_first=keep_only_first)
+                  keep_only_first=keep_only_first,
+                  remove_not_at_risk=remove_not_at_risk)
   return(out)
 }
 
@@ -49,12 +53,13 @@ as.data.table.simDT <- function(x, keep.rownames=FALSE, to,
 as.data.frame.simDT <- function(x, row.names=NULL, optional=FALSE,
                                 to, overlap=FALSE, target_event=NULL,
                                 keep_only_first=FALSE,
+                                remove_not_at_risk=FALSE,
                                 use_saved_states=x$save_states=="all",
                                 check_inputs=TRUE, ...) {
   out <- sim2data(sim=x, to=to, use_saved_states=use_saved_states,
                   as_data_frame=TRUE, check_inputs=check_inputs,
                   row.names=row.names, optional=optional, overlap=overlap,
                   target_event=target_event, keep_only_first=keep_only_first,
-                  ...)
+                  remove_not_at_risk=remove_not_at_risk, ...)
   return(out)
 }