+++ R/sim2start_stop_collapse.r
@@ -64,3 +64,48 @@ check_next_row_equal <- function(x) {
   shift_x <- data.table::shift(x, n=1, type="lead", fill=NA)
   return(rowSums(x==shift_x)==ncol(x))
 }
+
+## remove time not at-risk periods from start-stop data
+#' @importFrom data.table fifelse
+#' @importFrom data.table shift
+#' @importFrom data.table :=
+remove_not_at_risk <- function(data, duration, target_event, overlap) {
+
+  .last_event <- .id <- .start_in_event <- .stop_in_event <- start <- NULL
+
+  # if overlapping ones are supplied, simply re-transform to non-overlapping
+  # start-stop format, perform the transformation and re-add the + 1
+  if (overlap) {
+    data[, stop := stop - 1]
+  }
+
+  # get last event time
+  data[, .last_event := na_locf(fifelse(eval(parse(text=target_event))==TRUE,
+                                             stop, NA)), by=.id]
+  data[, .last_event := shift(.last_event, n=1), by=.id]
+  data[is.na(.last_event), .last_event := .Machine$integer.max]
+
+  # check if start is during an event
+  data[, .start_in_event := start > .last_event &
+         start < (.last_event + duration)]
+
+  # check if stop is during an event
+  data[, .stop_in_event := stop > .last_event &
+         stop < (.last_event + duration)]
+
+  # remove row if it is entirely during an event
+  data <- data[!(.start_in_event==TRUE & .stop_in_event==TRUE)]
+
+  # change start if only partially in there
+  data[.start_in_event==TRUE, start := .last_event + duration]
+
+  if (overlap) {
+    data[, stop := stop + 1]
+  }
+
+  data[, .last_event := NULL]
+  data[, .start_in_event := NULL]
+  data[, .stop_in_event := NULL]
+
+  return(data)
+}