+++ tests/testthat/test_sim2start_stop.all.r
@@ -20,6 +20,15 @@ test_that("outcome centric with event_duration > 1", {
   out2 <- sim2data(sim, to="start_stop", use_saved_states=TRUE,
                    target_event="sickness")
   expect_equal(out1, out2)
+
+  # less rows with remove_not_at_risk=TRUE
+  out3 <- sim2data(sim, to="start_stop", use_saved_states=FALSE,
+                   target_event="sickness", remove_not_at_risk=TRUE)
+  out4 <- sim2data(sim, to="start_stop", use_saved_states=TRUE,
+                   target_event="sickness", remove_not_at_risk=TRUE)
+  expect_equal(out3, out4)
+  expect_true(nrow(out3) < nrow(out1))
+  expect_true(nrow(out4) < nrow(out2))
 })
 
 test_that("outcome centric with event_duration == 1", {
@@ -43,4 +52,12 @@ test_that("outcome centric with event_duration == 1", {
   out2 <- sim2data(sim, to="start_stop", use_saved_states=TRUE,
                    target_event="sickness")
   expect_equal(out1, out2)
+
+  # adding remove_not_at_risk=TRUE changes nothing
+  out3 <- sim2data(sim, to="start_stop", use_saved_states=FALSE,
+                   target_event="sickness", remove_not_at_risk=TRUE)
+  out4 <- sim2data(sim, to="start_stop", use_saved_states=TRUE,
+                   target_event="sickness", remove_not_at_risk=TRUE)
+  expect_equal(out1, out3)
+  expect_equal(out2, out4)
 })